Test find_libpython
